Make the plugin work when in archived .sublime-package state
This requires a few workarounds but I also decided to assume that "Package/..."
paths are the only accepted for color schemes. This should very rarely result
in issues if at all and simplifies the path math a lot.
Furthermore, work around a bug in ST which prevents me from creating a dir
named `module_path` in the Packages dir. See SublimeText/Issues#55.