Attempt to fix crashes when dragging tabs
This only occurs in multi-group layouts and doesn't seem to follow a
pattern, or at least I couldn't find one. Adding a timeout (only for
ST2) seems to make that go away. With the way we dim schemes now this
shouldn't create flittering either.

The next release will probably be the last for ST2.